+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.4.25
 
+- Removed default() derive from TextureCache, as it needs to create a TTL Context from SDL2.
 - `TextureStore` changes:
   - TextWidget